refactor: Swagger 레이어 생성 후 분리 (#95)
