+++ src/main/java/com/codenear/butterfly/auth/presentation/OauthController.java
@@ -2,6 +2,7 @@ package com.codenear.butterfly.auth.presentation;
 
 import com.codenear.butterfly.auth.application.OauthService;
 import com.codenear.butterfly.auth.domain.dto.OauthDTO;
+import com.codenear.butterfly.auth.presentation.swagger.OauthControllerSwagger;
 import com.codenear.butterfly.global.dto.ResponseDTO;
 import com.codenear.butterfly.global.util.ResponseUtil;
 import jakarta.servlet.http.HttpServletResponse;
